feat: :sparkles: add auto-remove content-script logic and auto-insert on popup init
add auto-remove logic for the content-script when the context becomes invalidated; the check is every 10s. In popup now if it does not find a content-script it will try to add it before showing an error; the code asumes there will not be a content-script already present and active in the page.